# RamSoft.CodingTest.API/bin/Release/netcoreapp3.1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "RamSoft API",
    "version": "v1"
  },
  "paths": {
    "/ScrumTask": {
      "get": {
        "tags": [
          "Task"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/ScrumTask/{id}": {
      "get": {
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/ScrumTask/Add": {
      "post": {
        "tags": [
          "Task"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ScrumTask"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/ScrumTask/Update": {
      "put": {
        "tags": [
          "Task"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ScrumTask"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/ScrumTask/Delete/{id}": {
      "delete": {
        "tags": [
          "Task"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/ScrumTask/exception": {
      "get": {
        "tags": [
          "Task"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    }
  },
  "definitions": {
    "ScrumTask": {
      "required": [
        "description",
        "title"
      ],
      "type": "object",
      "properties": {
        "taskId": {
          "format": "int32",
          "type": "integer"
        },
        "title": {
          "maxLength": 250,
          "minLength": 1,
          "type": "string"
        },
        "description": {
          "minLength": 1,
          "type": "string"
        },
        "deadline": {
          "format": "date-time",
          "type": "string"
        },
        "originalEstimate": {
          "format": "int32",
          "type": "integer"
        },
        "completedHours": {
          "format": "int32",
          "type": "integer"
        },
        "statusName": {
          "type": "string"
        },
        "statusId": {
          "format": "int32",
          "type": "integer"
        },
        "status": {
          "$ref": "#/definitions/Status"
        }
      }
    },
    "Status": {
      "type": "object",
      "properties": {
        "statusId": {
          "format": "int32",
          "type": "integer"
        },
        "statusName": {
          "type": "string"
        }
      }
    }
  }
}